refactor(catalog): extract shared CatalogTypeNameFormatter
FormatTypeName (Nullable-unwrap + generic-arity strip + primitive-to-keyword
map, 29 LOC) was byte-identical in PromptParameterIndex and ToolParameterIndex,
both in Host.Stdio/Catalog/. Hoist into a shared internal CatalogTypeNameFormatter
in the same folder; both index builders delegate to it. Behavior unchanged.
Add a focused regression test pinning the label contract (primitives → keywords,
nullable value type → "T?", generic → "List<string>"/"Dictionary<string, int>").

Closes catalog-formattypename-dedup.
