+++ ai_docs/backlog.md
@@ -3,7 +3,7 @@
 <!-- purpose: Open work only; contract for agents syncing backlog on ship. -->
 <!-- scope: in-repo -->
 
-**updated_at: 2026-05-30T20:33:53Z**
+**updated_at: 2026-05-30T20:49:15Z**
 
 ## Agent contract
 
@@ -65,7 +65,6 @@
 | `swallow-fork-kill-empty-catch` | Low | none | Empty `catch {}` on the fork-restore `process.Kill(entireProcessTree: true)` swallows kill failures with no comment marking it a deliberate boundary. Add a one-line boundary comment or debug-log so the swallow is documented (matches the deliberate-boundary pattern already used in `WorkspaceTools.cs`). Anchors: `src/RoslynMcp.Host.Stdio/Tools/ValidationBundleTools.cs:355`. Regression test shape: comment-only; confirm build/`verify-docs` stay green. Evidence: doc-audit bad-code-surfacing 2026-05-28. [type: defect] |
 | `workspace-rootdir-vestigial-ternary` | Low | none | In `WaitForStableRestoreArtifactsAsync` the `EndsWith(".csproj") ? Path.GetDirectoryName(fullPath) : Path.GetDirectoryName(fullPath)` ternary has two identical branches, so the `.csproj` check is dead logic. Collapse to a single `Path.GetDirectoryName(fullPath)` call. Anchors: `src/RoslynMcp.Roslyn/Services/WorkspaceManager.cs:985`. Regression test shape: existing restore-artifact stability tests stay green after the collapse. Evidence: doc-audit bad-code-surfacing 2026-05-28. [type: defect] |
 | `firewallanalyzer-p3-polish-bundle-2026-05-16` | Low | none | [gh #769](https://github.com/darylmcd/Roslyn-Backed-MCP/issues/769) — bundled P3 polish list (10 cosmetic/UX findings + §14 improvement suggestions) from 2026-05-16 firewallanalyzer surface-test run against server v1.38.1. Sub-items §13.19-13.30: `find_duplicate_helpers` framework-wrapper false positives; `document_symbols`/`symbol_info` record-vs-class kind disagreement; `compile_check` file-filter scopes diagnostics but not compilation; `get_nuget_dependencies` returns literal `"centrally-managed"` under CPM; `get_cohesion_metrics` `lifecyclePattern`/`recommendation` always null; `add_pragma_suppression` emits CRLF in LF file; `get_msbuild_properties` vs `workspace_reload` `OutputType` mismatch; `source_file_lines` marker off-by-one vs `get_source_text.totalLineCount`; `find_type_mutations` error template diverges from sibling tools; `dependency_inversion_preview` newline-before-comma formatting; `remove_*_preview` family throws `InvalidOperation` instead of empty-preview for absent items; `go_to_definition` off-identifier error message misleads. Anchors: `audit-reports/20260516T062913Z_firewallanalyzer_mcp-server-surface-test.md` §13.19-13.30, §14. Evidence: 2026-05-16 surface-test run, server v1.38.1+7b2c0b9. |
-| `catalog-formattypename-dedup` | Low | none | Identical `FormatTypeName` (29 LOC — Nullable-unwrap + generic-arity strip + primitive-to-keyword switch) duplicated in `PromptParameterIndex` and `ToolParameterIndex` (same `Catalog/` folder). Extract into a shared internal static helper (e.g. `Catalog/CatalogTypeNameFormatter`) and call from both index builders. Anchors: `src/RoslynMcp.Host.Stdio/Catalog/PromptParameterIndex.cs:84`, `src/RoslynMcp.Host.Stdio/Catalog/ToolParameterIndex.cs:109`. Regression test shape: existing catalog parameter-index tests stay green. Evidence: `find_duplicated_code` similarity=1.0 (semantic). Source: 2026-05-28 discovery-sweep refactor. [type: refactor] |
 | `symbol-base-enumeration-dedup` | Low | none | Three near-identical methods `EnumerateMethodBases`/`EnumeratePropertyBases`/`EnumerateEventBases` (~28 LOC each) in `SymbolServiceHelpers` differ only by symbol type (`IMethodSymbol`/`IPropertySymbol`/`IEventSymbol`) and the `Overridden*` accessor. Judgment-bound: C# has no common `IOverridableSymbol`, so weigh a generic collapse (selector delegate) vs keeping three explicit methods for readability before applying. Anchors: `src/RoslynMcp.Roslyn/Helpers/SymbolServiceHelpers.cs:65`. Regression test shape: existing find_base_members / member-hierarchy tests stay green for method, property, and event inputs. Evidence: `find_duplicated_code` similarity=1.0 (semantic). Source: 2026-05-28 discovery-sweep refactor. [type: refactor] |
 | `mcp-registry-submission` | Low | external: MCP Registry approval (maintainer submission action, not a code change) | Submit the prepared `.claude-plugin/server.json` (name `io.github.darylmcd/roslyn-mcp`, NuGet `Darylmcd.RoslynMcp`, dnx) to the public MCP Registry; on approval, flip the README badge from "submission pending" to the live registry URL and add the registry-aware install snippet. Repo-side prep is already shipped (server.json draft via #502 + the registry-readiness check); only the external submit + badge-flip remain. Anchors: `.claude-plugin/server.json`, `README.md:4` (badge), `README.md:139`. Evidence: roadmap names registry publication as future direction (`docs/roadmap.md` § Claude Code Plugin Distribution); README badge still "submission pending"; no prior backlog row. Source: 2026-05-28 discovery-sweep roadmap-reconcile. |
 | `symbolrefactor-static-regex-generatedregex` | Low | none | Two compile-time-constant regex patterns are rebuilt via `new Regex(...)` on every call (no `RegexOptions.Compiled`, not source-generated) on the `split_service_with_di_preview` DI-rewrite path. Convert both to `[GeneratedRegex]` source-generated partials (or `static readonly Regex` fields). Cold/interactive path — payoff is hygiene + SYSLIB1045 compliance, not throughput. Anchors: `src/RoslynMcp.Roslyn/Services/SymbolRefactorService.cs:557`, `src/RoslynMcp.Roslyn/Services/SymbolRefactorService.cs:640`. Regression test shape: existing split-service-with-DI preview tests stay green (identical match behavior). Evidence: 2026-05-29 memoization audit; both patterns are non-interpolated string literals, `SemanticGrepService.cs:68` correctly excluded (user-supplied pattern, already `Compiled`). [type: refactor] |
